requirements.txt
gevent==20.6.2
requests==2.24.0
websocket-client==0.57.0